+++ R/mirt.R
@@ -193,6 +193,15 @@
 #' @param method a character object specifying the estimation algorithm to be used. The default is \code{'EM'},
 #'   for the standard EM algorithm with fixed quadrature. The option \code{'MHRM'} may also be passed to
 #'   use the MH-RM algorithm
+#' @param optimizer a character indicating which optimizer to use. By default, the EM algorithm 
+#'   will use the \code{'BFGS'} when there are no upper and lower bounds, and \code{'L-BFGS-B'}
+#'   when there are. Other options include the Newton-Raphson (\code{'NR'}), which often will
+#'   be more efficient than the \code{'BFGS'} but often not as stable for more complex models
+#'   (such as the nominal or nested logit models) or if poor start values are provided, 
+#'   and does not support upper and lower bound constraints. 
+#'   As well, the \code{'Nelder-Mead'} and \code{'SANN'} 
+#'   estimators may also be used, though their routine use generally is not required. 
+#'   The MH-RM algorithm uses the \code{'NR'} by default, and currently cannot be changed
 #' @param grsm.block an optional numeric vector indicating where the blocking should occur when using
 #'   the grsm, NA represents items that do not belong to the grsm block (other items that may be estimated
 #'   in the test data). For example, to specify two blocks of 3 with a 2PL item for the last item:
@@ -331,11 +340,6 @@
 #'          item factor analysis (\code{mixedmirt()} requires additional values for random effect).
 #'          If null, these values are determined internally, attempting to make the draws between
 #'          .1 and .4}
-#'     \item{Moptim}{Choose which optimizer to use. By default the EM algorithm with use the
-#'          \code{'BFGS'} when there are no upper and lower bounds, and \code{'L-BFGS-B'} when there
-#'          are. The \code{'Nelder-Mead'} and \code{'SANN'} estimators may also be used, though
-#'          their routine use generally is not required. The MH-RM algorithm uses the Newton-Raphson by
-#'          default, and currently cannot be changed}
 #'     \item{parallel}{logical; use the parallel cluster defined by \code{\link{mirtCluster}}?
 #'          Default is TRUE}
 #'   }
@@ -729,7 +733,7 @@
 #'
 #' }
 mirt <- function(data, model, itemtype = NULL, guess = 0, upper = 1, SE = FALSE, SE.type = 'crossprod',
-                 method = 'EM', pars = NULL, constrain = NULL, parprior = NULL,
+                 method = 'EM', optimizer = NULL, pars = NULL, constrain = NULL, parprior = NULL,
                  calcNull = TRUE, draws = 5000, survey.weights = NULL, rotate = 'oblimin', 
                  Target = NaN, quadpts = NULL, TOL = NULL, grsm.block = NULL, 
                  key = NULL, nominal.highlow = NULL, large = FALSE, GenRandomPars = FALSE,
@@ -743,7 +747,8 @@ mirt <- function(data, model, itemtype = NULL, guess = 0, upper = 1, SE = FALSE,
                       technical=technical, verbose=verbose, survey.weights=survey.weights,
                       calcNull=calcNull, SE.type=SE.type, large=large, key=key,
                       nominal.highlow=nominal.highlow, accellerate=accelerate, draws=draws,
-                      empiricalhist=empiricalhist, GenRandomPars=GenRandomPars, ...)
+                      empiricalhist=empiricalhist, GenRandomPars=GenRandomPars, 
+                      optimizer=optimizer, ...)
     if(is(mod, 'ExploratoryClass') || is(mod, 'ConfirmatoryClass'))
         mod@Call <- Call
     return(mod)